+++ src/Chat/Chat.module.css
@@ -1,15 +1,3 @@
-.popup {
-  position: fixed;
-  bottom: 0;
-  right: 0;
-  margin: 0 10px 10px 0;
-  box-shadow: rgba(0, 0, 0, 0.3) 0px 19px 38px,
-    rgba(0, 0, 0, 0.22) 0px 15px 12px;
-}
-.comments {
-  overflow-y: scroll;
-  height: 100%;
-}
 .container {
   background: #443a84;
   position: fixed;
@@ -32,24 +20,18 @@
   opacity: 0.8;
   background-color: black;
 }
-.closeIconContainer {
+.closeIcon {
   display: flex;
   justify-content: flex-end;
+  padding: 12px;
 }
-.closeIcon {
-  color: white;
-  display: flex;
-  align-items: center;
-  padding: 6px;
-  justify-content: center;
-}
-.messageEnd {
-  float: left;
-  clear: both;
-}
-.user {
-  color: greenyellow;
-  margin-bottom: 2px;
+.popup {
+  position: fixed;
+  bottom: 0;
+  right: 0;
+  margin: 0 10px 10px 0;
+  box-shadow: rgba(0, 0, 0, 0.3) 0px 19px 38px,
+    rgba(0, 0, 0, 0.22) 0px 15px 12px;
 }
 .alert {
   position: absolute;
@@ -59,56 +41,4 @@
   height: 10px;
   border-radius: 50%;
   background-color: red;
-}
-.comment {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  background-color: #02b7dd;
-  margin: 0 38px 16px 16px;
-  padding: 10px;
-  border-radius: 4px;
-}
-.comment:after {
-  content: "";
-  position: absolute;
-  left: 0;
-  top: 80%;
-  width: 0;
-  height: 0;
-  border: 10px solid transparent;
-  border-right-color: #02b7dd;
-  border-left: 0;
-  border-bottom: 0;
-  margin-top: -10px;
-  margin-left: -10px;
-}
-.myComment {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  color: white;
-  background-color: #02b7dd;
-  margin-bottom: 16px;
-  padding: 10px 10px 12px 10px;
-  margin-right: 15px;
-  margin-left: 38px;
-  border-radius: 4px;
-  border-width: 1px;
-}
-.myComment:after {
-  content: "";
-  position: absolute;
-  right: 0;
-  top: 80%;
-  width: 0;
-  height: 0;
-  border: 10px solid transparent;
-  border-left-color: #02b7dd;
-  border-right: 0;
-  border-bottom: 0;
-  margin-top: -10px;
-  margin-right: -10px;
 }
